+++ tests/scriptlets/trusted-replace-node-text.test.js
@@ -22,9 +22,21 @@ const afterEach = () => {
 
 module(name, { beforeEach, afterEach });
 
+let policy;
+if (window.trustedTypes) {
+    policy = window.trustedTypes.createPolicy('myEscapePolicy', {
+        createHTML: (string) => string,
+        createScript: (string) => string,
+    });
+}
+
 const addNode = (nodeName, textContent) => {
     const node = document.createElement(nodeName);
-    node.textContent = textContent;
+    if (nodeName === 'script' && policy) {
+        node.textContent = policy.createScript(textContent);
+    } else {
+        node.textContent = textContent;
+    }
     elements.push(node);
     document.body.appendChild(node);
     return node;
@@ -86,6 +98,44 @@ test('using matchers as regexes', (assert) => {
         assert.strictEqual(safeNodeBefore.textContent, text, 'non-matched node should not be affected');
 
         assert.strictEqual(window.hit, 'FIRED', 'hit function should fire');
+        done();
+    }, 1);
+});
+
+test('CSP require-trusted-types-for "script" - replace script content', (assert) => {
+    const done = assert.async();
+
+    // QUnit uses innerHTML to render tests, so we need to override it and add trusted types policy
+    const nativeInnerHTMLSet = Object.getOwnPropertyDescriptor(Element.prototype, 'innerHTML').set;
+    Object.defineProperty(Element.prototype, 'innerHTML', {
+        set(html) {
+            if (policy) {
+                const sanitizedHTML = policy.createHTML(html);
+                return nativeInnerHTMLSet.call(this, sanitizedHTML);
+            }
+            return nativeInnerHTMLSet.call(this, html);
+        },
+    });
+
+    const meta = document.createElement('meta');
+    document.head.appendChild(meta);
+    meta.setAttribute('http-equiv', 'Content-Security-Policy');
+    meta.setAttribute('content', "require-trusted-types-for 'script'");
+
+    const nodeName = 'script';
+    const textMatch = 'window.showAds';
+    const pattern = 'window.showAds = true';
+    const replacement = 'window.showAds = false';
+
+    const expectedText = 'window.showAds = false;';
+
+    runScriptlet(name, [nodeName, textMatch, pattern, replacement]);
+
+    const nodeAfter = addNode('script', 'window.showAds = true;');
+    setTimeout(() => {
+        assert.strictEqual(nodeAfter.textContent, expectedText, 'text content should be modified');
+        assert.strictEqual(window.hit, 'FIRED', 'hit function should fire');
+        clearGlobalProps('showAds');
         done();
     }, 1);
 });